Make most classes Serializable and added the DoubleValue type
This project is intended to be used as a library for apps compatible with clustered systems such as Spark.
Hence, most classes need to be serializable so that copies of object can be sent over the network to machines of the cluster.

Currently does not work for predicates, but works for basic rules.

* Added a SerializeAutomaton utility function in the Rule Manager for exporting a rule automaton to a file.